requirements.txt
fastapi
uvicorn
python-dotenv
langchain-community
langchain-core 
pypdf
numpy
sentence-transformers
chromadb
google-genai
pydantic